+++ src/shared/components/ThemeProvider.js
@@ -1,15 +1,2 @@
-"use client";
-
-import { useEffect } from "react";
-import useThemeStore from "@/store/themeStore";
-
-export function ThemeProvider({ children }) {
-  const { initTheme } = useThemeStore();
-
-  useEffect(() => {
-    initTheme();
-  }, [initTheme]);
-
-  return <>{children}</>;
-}
-
+export { default } from "./layout/ThemeProvider";
+export * from "./layout/ThemeProvider";